Moving global functions to a helper module
