+++ src/atomate2/jdftx/io/joutstructures.py
@@ -1,7 +1,9 @@
 from dataclasses import dataclass
 from typing import Any
 
-from atomate2.jdftx.io.joutstructure import JOutStructure, correct_iter_type, is_lowdin_start_line
+from jdftx.io.joutstructure_helpers import correct_iter_type, is_lowdin_start_line
+
+from atomate2.jdftx.io.joutstructure import JOutStructure
 
 
 def get_start_idx(out_slice: list[str], out_slice_start_flag: str = "-------- Electronic minimization -----------") -> int:
